точка входа, разделение участка кода по модулям
